Rewriting is done, also found the problem of the commandbug.

+++ Project/core/src/com/ives/relative/entities/components/network/NetworkC.java
@@ -7,13 +7,13 @@ import com.ives.relative.managers.NetworkManager;
  * Created by Ives on 13/12/2014.
  */
 public class NetworkC extends Component implements Networkable {
-    public long id;
+    public int id;
     public NetworkManager.Type type;
 
     public NetworkC() {
     }
 
-    public NetworkC(long id, NetworkManager.Type type) {
+    public NetworkC(int id, NetworkManager.Type type) {
         this.id = id;
         this.type = type;
     }
