+++ Project/core/src/com/ives/relative/network/packets/UpdatePacket.java
@@ -4,12 +4,12 @@ package com.ives.relative.network.packets;
  * Created by Ives on 14/12/2014.
  */
 public abstract class UpdatePacket extends BasePacket {
-    public long entityID;
+    public int entityID;
 
     public UpdatePacket() {
     }
 
-    public UpdatePacket(int sequence, long entityID) {
+    public UpdatePacket(int sequence, int entityID) {
         super(sequence);
         this.entityID = entityID;
     }
